+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Not yet defined
+- Fix reconnect AWS IOT MQTT broker by publishing messages with QOS=1 (At least once)
+- Add AWS Broker status sensor
+
 ## v0.0.9
 - Removed time from cleaning mode select HA component (select, translations and services)
 - Moved the off_states of switch to constants